+++ test/base/common/file/path.test.ts
