+++ test/base/common/file/stream.test.ts
