+++ test/base/common/file/uri.test.ts
